+++ mse-20190531/core/Models/CreateNacosMcpServerRequest.cs
@@ -17,6 +17,10 @@ namespace AlibabaCloud.SDK.Mse20190531.Models
         [Validation(Required=false)]
         public string AcceptLanguage { get; set; }
 
+        [NameInMap("EncryptToolSpec")]
+        [Validation(Required=false)]
+        public bool? EncryptToolSpec { get; set; }
+
         /// <summary>
         /// <b>Example:</b>
         /// <para>{&quot;type&quot;:&quot;REF&quot;,&quot;data&quot;:{&quot;namespaceId&quot;:&quot;public&quot;,&quot;groupName&quot;:&quot;mcp-endpoints&quot;,&quot;serviceName&quot;:&quot;mcp-demo&quot;}}</para>